Tests coverage is 100% on service files. Got rid of singleton pattern (unique instance "shared") in order to be able to delete private override init(). The latter was causing problems because of its access control level and singleton pattern here was unnecessary.
Coverage to 100% was also reached by covering the failure case in which the URL was bad, in all services files. This was done by breaking down all API URLs, and creating a parameter that would allow changes to the base URL's value, in order to break it and fall into the failure case.